requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==1.0.0
Django==4.0.5
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-on-heroku==1.1.2
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dotenv==0.21.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.10
whitenoise==6.2.0
